chatbot-web/requirements.txt
flask
google-generativeai
uuid
